requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-mongo
pandas
pymongo